recommender: fast unit vector
This patch adds fast_unit_vector/gen_uv to the Recommender API. This will be used for generating user unit-vectors, and may be used as an auxillary function for the database helper.

+++ cpp/common.h
@@ -3,4 +3,4 @@
 #include <cstdint>
 
 typedef uint64_t u64;
-typedef float f32;
+typedef double f64;
