+++ cpp/mathtools.h
@@ -2,8 +2,8 @@
 
 #pragma once
 
-f32 dotProduct(f32 v1[], f32 v2[], u64 len);
+f64 dotProduct(f64 v1[], f64 v2[], u64 len);
 
-f32 vectorMagnitude(f32 v[], u64 len);
+f64 vectorMagnitude(f64 v[], u64 len);
 
-void makeVectorUnitInPlace(f32 v[], u64 len);
+void makeVectorUnitInPlace(f64 v[], u64 len);
